+++ src/test/java/io/hhplus/tdd/ChargePointServiceTest.java
@@ -4,9 +4,9 @@ import io.hhplus.tdd.point.application.ChargePointService;
 import io.hhplus.tdd.point.application.PointService;
 import io.hhplus.tdd.point.domain.PointHistory;
 import io.hhplus.tdd.point.domain.UserPoint;
-import io.hhplus.tdd.point.infra.DefaultChargePointService;
-import io.hhplus.tdd.point.infra.FairReentrantLockChargePointService;
-import io.hhplus.tdd.point.infra.SynchronizedChargePointService;
+import io.hhplus.tdd.point.infra.service.point.DefaultChargePointService;
+import io.hhplus.tdd.point.infra.service.point.FairReentrantLockChargePointService;
+import io.hhplus.tdd.point.infra.service.point.SynchronizedChargePointService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
